requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.3.3
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
zipp==3.16.2
